KINETIS platform update for ChibiOS 18.2.x.
